lightweightnode/requirements.txt
flask
requests
ecdsa
pybloom-live
Charm-Crypto
